requirements.txt
opencv-python-headless
vidgear
yt_dlp
google-cloud-storage